+++ internal/parser/zz_ast_test.go
@@ -458,3 +458,19 @@ func TestTabExpr_Pos(t *testing.T) {
 	testutils.Equal(t, "column", col, 4)
 	testutils.Equal(t, "token", tok, "TAB")
 }
+
+func TestLeftExpr_Pos(t *testing.T) {
+	expr := &LeftExpr{
+		StrExpr: &StringLiteral{Value: "HELLO"},
+		LenExpr: &NumberLiteral{Value: 5},
+		Line:    20,
+		Column:  8,
+		Token:   "LEFT$",
+	}
+
+	line, col, tok := expr.Pos()
+
+	testutils.Equal(t, "line", line, 20)
+	testutils.Equal(t, "column", col, 8)
+	testutils.Equal(t, "token", tok, "LEFT$")
+}